% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityTools.R
\name{density_from_mass_abm}
\alias{density_from_mass_abm}
\title{Calculate the density from \% alcohol by mass}
\usage{
density_from_mass_abm(mass_abm, temperature)
}
\arguments{
\item{mass_abm}{Alcohol \% by mass (numeric)}

\item{temperature}{Temperature in degrees C (numeric)}
}
\value{
a density
}
\description{
Calculate the density from \% alcohol by mass
}
