% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityTools.R
\name{density_from_vol_abv}
\alias{density_from_vol_abv}
\title{Calculate density and alcohol mass fraction from alcohol \% by volume}
\usage{
density_from_vol_abv(vol_abv, temperature)
}
\arguments{
\item{vol_abv}{Alcohol \% by volume (numeric)}

\item{temperature}{temperature in degrees C (numeric)}
}
\value{
a list with density at 20 degrees and mass fraction
}
\description{
Calculate density and alcohol mass fraction from alcohol \% by volume
}
